% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_SIR.R
\name{CP_SIR}
\alias{CP_SIR}
\title{Counting-Process Based Sliced Inverse Regression}
\usage{
CP_SIR(x, y, censor, bw = silverman(1, length(y)))
}
\arguments{
\item{x}{A matrix (n x p) of features (continuous only).}

\item{y}{A numeric vector of observed survival times.}

\item{censor}{A numeric vector of censoring indicators (1 = event,
0 = censored).}

\item{bw}{Numeric; kernel bandwidth for nonparametric estimation
(one-dimensional). Default: Silverman's rule via \code{\link{silverman}}.}
}
\value{
A list with components:
\describe{
\item{\code{values}}{Eigenvalues of the estimation matrix.}
\item{\code{vectors}}{Estimated directions (columns), ordered by
eigenvalues.}
}
}
\description{
Fits the CP-SIR model for right-censored survival outcomes. This model is
correct only under strong assumptions, but since it requires only an SVD,
its solution is commonly used as the initial value in
\code{\link{orthoDr_surv}} optimization.
}
\examples{
# This is setting 1 in Sun et. al. (2017) with reduced sample size
library(MASS)
set.seed(1)
N = 200; P = 6
V=0.5^abs(outer(1:P, 1:P, "-"))
dataX = as.matrix(mvrnorm(N, mu=rep(0,P), Sigma=V))
failEDR = as.matrix(c(1, 0.5, 0, 0, 0, rep(0, P-5)))
censorEDR = as.matrix(c(0, 0, 0, 1, 1, rep(0, P-5)))
T = rexp(N, exp(dataX \%*\% failEDR))
C = rexp(N, exp(dataX \%*\% censorEDR - 1))
ndr = 1
Y = pmin(T, C)
Censor = (T < C)

# fit the model
cpsir.fit = CP_SIR(dataX, Y, Censor)
distance(failEDR, cpsir.fit$vectors[, 1:ndr, drop = FALSE], "dist")
}
\references{
Sun, Q., Zhu, R., Wang, T. and Zeng, D. (2017). Counting process based
dimension reduction method for censored outcomes.
\emph{arXiv preprint} \doi{10.48550/arXiv.1704.05046}.
}
\seealso{
\code{\link{orthoDr_surv}}, \code{\link{silverman}}
}
