% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dosepred}
\alias{dosepred}
\title{The prediction function for the personalized direct learning dose model}
\usage{
dosepred(B, X, X_test, bw, W)
}
\arguments{
\item{B}{A matrix of the parameters \code{B}, the columns are subject to the orthogonality constraint}

\item{X}{The covariate matrix}

\item{X_test}{The test covariate matrix}

\item{bw}{A Kernel bandwidth, assuming each variable have unit variance}

\item{w}{The kernel ridge regression coefficient}
}
\value{
The predicted dose
}
\description{
Predict the fitted dose from the direct learning dose model
}
\keyword{internal}
