% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KernelDist_cross}
\alias{KernelDist_cross}
\title{KernelDist_cross}
\usage{
KernelDist_cross(TestX, X)
}
\arguments{
\item{TestX}{testing data}

\item{X}{training data}
}
\description{
Calculate the kernel distance between testing data and training data
}
\keyword{internal}
