% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{local_solver}
\alias{local_solver}
\title{local semi regression solver \code{C++} function}
\usage{
local_solver(
  B,
  X,
  Y,
  bw,
  rho,
  eta,
  gamma,
  tau,
  epsilon,
  btol,
  ftol,
  gtol,
  maxitr,
  verbose,
  ncore
)
}
\arguments{
\item{B}{A matrix of the parameters \code{B}, the columns are subject to the orthogonality constraint}

\item{X}{A matrix of the parameters \code{X}}

\item{Y}{A matrix of the parameters \code{Y}}

\item{bw}{Kernel bandwidth for X}

\item{rho}{(don't change) Parameter for control the linear approximation in line search}

\item{eta}{(don't change) Factor for decreasing the step size in the backtracking line search}

\item{gamma}{(don't change) Parameter for updating C by Zhang and Hager (2004)}

\item{tau}{(don't change) Step size for updating}

\item{epsilon}{(don't change) Parameter for apprximating numerical gradient, if \code{g} is not given.}

\item{btol}{(don't change) The \code{$B$} parameter tolerance level}

\item{ftol}{(don't change) Functional value tolerance level}

\item{gtol}{(don't change) Gradient tolerance level}

\item{maxitr}{Maximum number of iterations}

\item{verbose}{Should information be displayed}
}
\description{
Sovling the local semiparametric estimating equations. This is an internal function and should not be called directly.
}
\references{
Ma, Y., & Zhu, L. (2013). Efficient estimation in sufficient dimension reduction.
Annals of statistics, 41(1), 250.
DOI: \doi{10.1214/12-AOS1072}
}
\keyword{internal}
