% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoDr-package.R
\docType{package}
\name{orthoDr-package}
\alias{orthoDr}
\alias{orthoDr-package}
\title{orthoDr: Semi-Parametric Dimension Reduction Models Using Orthogonality Constrained Optimization}
\description{
Provides semi-parametric dimension reduction methods using
orthogonality constrained optimization. Includes regression, survival,
and personalized dose models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqzhu-aide/orthoDr}
  \item Report bugs at \url{https://github.com/rqzhu-aide/orthoDr/issues}
}

}
\author{
\strong{Maintainer}: Ruoqing Zhu \email{teazrq@gmail.com} (\href{https://orcid.org/0000-0002-0753-5716}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ruoqing Zhu \email{teazrq@gmail.com} (\href{https://orcid.org/0000-0002-0753-5716}{ORCID}) [copyright holder]
  \item Ruilin Zhao
  \item Wenzhuo Zhou
}

Other contributors:
\itemize{
  \item Jiyang Zhang [contributor]
  \item Peng Xu [contributor]
}

}
\keyword{internal}
