% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoDr_pt.r
\name{orthoDr_pm}
\alias{orthoDr_pm}
\title{orthoDr_pm model}
\usage{
orthoDr_pm(x, a, r, ndr = ndr, B.initial = NULL, bw = NULL,
  lambda = 0.1, K = 20, method = c("direct_kernel", "semi_svm"),
  keep.data = FALSE, control = list(), maxitr = 500,
  verbose = FALSE, ncore = 0)
}
\arguments{
\item{x}{A matrix or data.frame for features (continuous only).}

\item{a}{A vector of observed dose}

\item{r}{A vector of observed reward}

\item{ndr}{A dimension structure}

\item{B.initial}{Initial \code{B} values. Will use the counting process based SIR model \link[orthoDr]{CP_SIR} as the initial if leaving as \code{NULL}.
If specified, must be a matrix with \code{ncol(x)} rows and \code{ndr} columns. Will be processed by Gram-Schmidt if not orthogonal}

\item{bw}{A Kernel bandwidth, assuming each variables have unit variance}

\item{lambda}{A GCV penalty for the kernel ridge regression}

\item{K}{A number of grids in the range of dose}

\item{method}{A method the user will implement}

\item{keep.data}{Should the original data be kept for prediction}

\item{control}{A list of tuning variables for optimization. \code{epsilon} is the size for numerically appriximating the gradient. For others, see Wen and Yin (2013).}

\item{maxitr}{Maximum number of iterations}

\item{verbose}{Should information be displayed}

\item{ncore}{the number of cores for parallel computing}
}
\value{
A \code{orthoDr} object; a list consisting of
\item{B}{The optimal \code{B} value}
\item{fn}{The final functional value}
\item{itr}{The number of iterations}
\item{converge}{convergence code}
}
\description{
The "Direct Learning & Pseudo-direct Learning" Method for personalized medicine.
}
\references{
Zhou, W., Zhu, R. "A Parsimonious Personalized Dose Model vis Dimension Reduction." (2018)
\url{https://arxiv.org/abs/1802.06156}.

Wen, Z. and Yin, W., "A feasible method for optimization with orthogonality constraints." Mathematical Programming 142.1-2 (2013): 397-434.
DOI: \url{https://doi.org/10.1007/s10107-012-0584-1}
}
