% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoDr-package.R
\docType{package}
\name{orthoDrLite-package}
\alias{orthoDrLite}
\alias{orthoDrLite-package}
\title{orthoDrLite: Semi-Parametric Dimension Reduction Models Using Orthogonality
        Constrained Optimization}
\description{
Utilize an orthogonality constrained optimization algorithm of
    Wen & Yin (2013) <DOI:10.1007/s10107-012-0584-1> to solve a variety of
    dimension reduction problems in the semiparametric framework, such as
    Ma & Zhu (2012) <DOI:10.1080/01621459.2011.646925>, Ma & Zhu (2013)
    <DOI:10.1214/12-AOS1072>, Sun, Zhu, Wang & Zeng (2017) <arXiv:1704.05046>
    and Zhou & Zhu (2018+) <arXiv:1802.06156>. It also serves as a general
    purpose optimization solver for problems with orthogonality constraints.
    Parallel computing for approximating the gradient is enabled
    through 'OpenMP'. Forked from orthoDr and removed the dependencies on rgl.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Displayr/orthoDrLite}
  \item Report bugs at \url{https://github.com/Displayr/orthoDrLite/issues}
}

}
\author{
\strong{Maintainer}: Ruoqing Zhu \email{teazrq@gmail.com} (\href{https://orcid.org/0000-0002-0753-5716}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ruilin Zhao \email{rzhao15@seas.upenn.edu} [copyright holder]
  \item Jiyang Zhang \email{jiyangz2@illinois.edu} [copyright holder]
  \item Wenzhuo Zhou \email{wenzhuo3@illinois.edu} [copyright holder]
  \item Peng Xu \email{px2132@columbia.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [contributor]
}

}
\keyword{internal}
