% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_SAVE.r
\name{P_SAVE}
\alias{P_SAVE}
\title{Partial Sliced Averaged Variance Estimation}
\usage{
P_SAVE(x, a, r, ndr = 2, nslices0 = 2)
}
\arguments{
\item{x}{A matrix for features (continuous only).}

\item{a}{A vector of observed dose levels (continuous only).}

\item{r}{A vector of reward (outcome).}

\item{ndr}{The central subspace dimension}

\item{nslices0}{Number of slides used for save}
}
\value{
A list consisting of
\item{vectors}{The basis of central subspace, ordered by eigenvalues}
}
\description{
The P_SAVE model. This model is correct only under very strong assumptions, the solution is used as the initial value in the orthoDr optimization.
}
\references{
Feng, Z., Wen, M.X, Yu, Z. and Zhu L. "On Partial Sufficient Dimension Reduction With Applications to Partially Linear Multi-Index Models" (2013)
\url{https://arxiv.org/abs/1704.05046} .
}
