% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_check.r
\name{param_check}
\alias{param_check}
\alias{control.check}
\title{param_check}
\usage{
control.check(control)
}
\arguments{
\item{control}{control parameters}
}
\description{
Check parameters
}
\keyword{internal}
