% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict_orthoDr_reg}
\alias{predict_orthoDr_reg}
\title{predict_orthoDr_reg}
\usage{
predict_orthoDr_reg(object, testx, ...)
}
\arguments{
\item{object}{fitted object}

\item{testx}{Testing data}

\item{...}{...}
}
\value{
The predicted object
}
\description{
Internal prediction function for regression models
}
\keyword{internal}
