% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.orthoDr}
\alias{print.orthoDr}
\title{print.orthoDr}
\usage{
\method{print}{orthoDr}(x, ...)
}
\arguments{
\item{x}{A fitted orthoDr object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisible NULL. Called for its side-effect (printing to console).
}
\description{
print.orthoDr
}
\examples{
# generate some survival data
N = 100; P = 4; dataX = matrix(rnorm(N*P), N, P)
Y = exp(-1 + dataX[,1] + rnorm(N))
Censor = rbinom(N, 1, 0.8)

# fit the model
orthoDr_surv(dataX, Y, Censor, ndr = 1, method = "dm")
}
\keyword{internal}
