% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silverman.R
\name{silverman}
\alias{silverman}
\title{A simple Silverman bandwidth formula}
\usage{
silverman(d, n)
}
\arguments{
\item{d}{Integer; number of dimensions.}

\item{n}{Integer; number of observations.}
}
\value{
A numeric scalar: the Silverman bandwidth.
}
\description{
Compute the Silverman rule-of-thumb bandwidth for kernel density estimation.
}
\examples{
silverman(1, 300)
}
