% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skcm.r
\docType{data}
\name{skcm.clinical}
\alias{skcm.clinical}
\title{skcm.clinical}
\format{Contains 469 subjects with 156 failured. Each row contains one subject, subject ID is indicated by row name. Variables include \code{Time}, \code{Censor}, \code{Gender} and \code{Age}. \code{Age} has 8 missing values.}
\usage{
skcm.clinical
}
\description{
The clinical varaibles of the SKCM dataset. The original data was obtained from The Cancer Genome Atlas (TCGA).
}
\references{
\url{https://cancergenome.nih.gov/}
}
\keyword{skcm.clinical}
