% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skcm.r
\docType{data}
\name{skcm.melgene}
\alias{skcm.melgene}
\title{skcm.melgene}
\format{Each row contains one subject, subject ID is indicated by row name. Gene names in the columns. The columns are scaled.}
\usage{
skcm.melgene
}
\description{
The expression of top 20 genes of cutaneous melanoma literature based on Athanasiadis, et al. 2014.
}
\references{
Athanasiadis, Emmanouil I., et al. "A Web-based database of genetic association studies in cutaneous melanoma enhanced with network-driven data exploration tools." Database 2014 (2014).
}
\keyword{skcm.melgene}
