% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Dosepred}
\alias{Dosepred}
\title{Dosepred}
\usage{
Dosepred(B, X, X_test, bw, W)
}
\arguments{
\item{B}{A matrix of the parameters \code{B}, the columns are subject to the orthogonality constraint}

\item{X}{The covariate matrix}

\item{X_test}{Test x}

\item{bw}{A Kernel bandwidth, assuming each variable have unit variance}

\item{w}{The kernel ridge regression w}
}
\value{
The Dose.
}
\description{
calculate the predicted dose
}
\keyword{internal}
