% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoDr-package.R
\docType{package}
\name{orthoDr-package}
\alias{orthoDr}
\alias{orthoDr-package}
\title{orthoDr: Semi-Parametric Dimension Reduction Models Using Orthogonality Constrained Optimization}
\description{
Utilize an orthogonality constrained optimization algorithm of Wen & Yin (2013) \doi{10.1007/s10107-012-0584-1} to solve a variety of dimension reduction problems in the semiparametric framework, such as Ma & Zhu (2012) \doi{10.1080/01621459.2011.646925}, Ma & Zhu (2013) \doi{10.1214/12-AOS1072}, Sun, Zhu, Wang & Zeng (2019) \doi{10.1093/biomet/asy064} and Zhou, Zhu & Zeng (2021) \doi{10.1093/biomet/asaa087}. The package also implements some existing dimension reduction methods such as hMave by Xia, Zhang, & Xu (2010) \doi{10.1198/jasa.2009.tm09372} and partial SAVE by Feng, Wen & Zhu (2013) \doi{10.1080/01621459.2012.746065}. It also serves as a general purpose optimization solver for problems with orthogonality constraints, i.e., in Stiefel manifold. Parallel computing for approximating the gradient is enabled through 'OpenMP'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/teazrq/orthoDr}
  \item Report bugs at \url{https://github.com/teazrq/orthoDr/issues}
}

}
\author{
\strong{Maintainer}: Ruoqing Zhu \email{teazrq@gmail.com} (\href{https://orcid.org/0000-0002-0753-5716}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ruilin Zhao \email{rzhao15@seas.upenn.edu} [copyright holder]
  \item Jiyang Zhang \email{jiyangz2@illinois.edu} [copyright holder]
  \item Wenzhuo Zhou \email{wenzhuo3@illinois.edu} [copyright holder]
  \item Peng Xu \email{px2132@columbia.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [contributor]
}

}
\keyword{internal}
