% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSAVE.r
\name{pSAVE}
\alias{pSAVE}
\title{Partial Sliced Averaged Variance Estimation}
\usage{
pSAVE(x, a, r, ndr = 2, nslices0 = 2)
}
\arguments{
\item{x}{A \code{matrix} for features (continuous only).}

\item{a}{A \code{vector} of observed dose levels (continuous only).}

\item{r}{A \code{vector} of reward (outcome).}

\item{ndr}{The dimension structure}

\item{nslices0}{Number of slides used for save}
}
\value{
A \code{list} consisting of:

\item{vectors}{The basis of central subspace, ordered by eigenvalues}
}
\description{
The partial-SAVE model. This model is correct only under very strong
assumptions, the solution is used as the initial value in the \code{orthoDr}
optimization.
}
\references{
Feng, Z., Wen, M.X, Yu, Z. and Zhu L. "On Partial Sufficient Dimension Reduction With Applications to Partially Linear Multi-Index Models" (2013)
\url{https://arxiv.org/abs/1704.05046} .
}
