% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silverman.r
\name{silverman}
\alias{silverman}
\title{Silverman's rule of thumb}
\usage{
silverman(d, n)
}
\arguments{
\item{d}{Number of dimension}

\item{n}{Number of observation}
}
\value{
A simple bandwidth choice
}
\description{
A simple Silverman's rule of thumb bandwidth calculation.
}
\examples{
silverman(1, 300)
}
