% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skcm.r
\docType{data}
\name{skcm.clinical}
\alias{skcm.clinical}
\title{Skin Cutaneous Melanoma Data set}
\format{
Contains 469 subjects with 156 failures. Each row contains one subject,
subject ID is indicated by row name. Variables include:
\itemize{
\item \code{Time}
\item \code{Censor}
\item \code{Gender}
\item \code{Age}
}

\strong{Note:} \code{Age} has 8 missing values.
}
\usage{
skcm.clinical
}
\description{
The clinical variables of the SKCM dataset.
The original data was obtained from The Cancer Genome Atlas (TCGA).
}
\references{
\url{https://www.cancer.gov/about-nci/organization/ccg/research/structural-genomics/tcga}
}
\keyword{datasets}
